--- lib/bureaucrat/fields.rb
@@ -1,12 +1,0 @@
-Dir[File.expand_path('../fields/*.rb', __FILE__)].each { |file| require file }
-# TimeField
-# DateTimeField
-#class ImageField < FileField
-#end
-
-# URLField
-# TypedMultipleChoiceField < MultipleChoiceField
-# MultiValueField
-# FilePathField
-# SplitDateTimeField
-
