+++ lib/bureaucrat/fields/field.rb
@@ -1,3 +1,4 @@
+require 'bureaucrat'
 require 'bureaucrat/utils'
 require 'bureaucrat/validation_error'
 require 'bureaucrat/widgets/hidden_input'
