+++ lib/bureaucrat/formsets.rb
@@ -1,5 +1,6 @@
-require 'bureaucrat/fields'
-require 'bureaucrat/forms'
+require 'bureaucrat'
+require 'bureaucrat/fields/boolean_field'
+require 'bureaucrat/fields/integer_field'
 require 'bureaucrat/utils'
 require 'bureaucrat/widgets/hidden_input'
 
@@ -11,20 +12,18 @@ module Bureaucrat
     ORDERING_FIELD_NAME = :'ORDER'
     DELETION_FIELD_NAME = :'DELETE'
 
-    class ManagementForm < Forms::Form
+    class ManagementForm < Form
       include Fields
       include Widgets
 
-      field TOTAL_FORM_COUNT, IntegerField.new(widget: HiddenInput)
-      field INITIAL_FORM_COUNT, IntegerField.new(widget: HiddenInput)
-      field MAX_NUM_FORM_COUNT, IntegerField.new(widget: HiddenInput,
-                                                 required: false)
+      field TOTAL_FORM_COUNT, Fields::IntegerField.new(widget: Widgets::HiddenInput)
+      field INITIAL_FORM_COUNT, Fields::IntegerField.new(widget: Widgets::HiddenInput)
+      field MAX_NUM_FORM_COUNT, Fields::IntegerField.new(widget: Widgets::HiddenInput,
+                                                         required: false)
     end
 
     class BaseFormSet
       include Utils
-      include Fields
-      include Forms
 
       def self.default_prefix
         'form'
@@ -60,7 +59,7 @@ module Bureaucrat
         @auto_id = options.fetch(:auto_id, 'id_%s')
         @data = data || {}
         @initial = options[:initial]
-        @error_class = options.fetch(:error_class, ErrorList)
+        @error_class = options.fetch(:error_class, Fields::ErrorList)
         @errors = nil
         @non_form_errors = nil
 
@@ -259,7 +258,7 @@ module Bureaucrat
           form.fields[ORDERING_FIELD_NAME] = IntegerField.new(attrs)
         end
         if can_delete
-          field = BooleanField.new(label: 'Delete', required: false)
+          field = Fields::BooleanField.new(label: 'Delete', required: false)
           form.fields[DELETION_FIELD_NAME] = field
         end
       end
