+++ lib/bureaucrat/validators/base.rb
@@ -1,3 +1,5 @@
+require 'bureaucrat'
+
 module Bureaucrat
   module Validators
     class BaseValidator
