+++ test/validators/comma_separated_integer_list_test.rb
@@ -5,7 +5,7 @@ module TestValidators
   class Test_ValidateCommaSeparatedIntegerList < BureaucratTestCase
     def test_translated_message
       Bureaucrat::Validators::ValidateCommaSeparatedIntegerList.call(nil)
-    rescue ValidationError => e
+    rescue Bureaucrat::ValidationError => e
       assert_equal([I18n.t('bureaucrat.default_errors.validators.validate_comma_separated_integer_list')], e.messages)
     end
   end
