+++ test/widgets/file_input_test.rb
@@ -4,7 +4,7 @@ require 'bureaucrat/widgets/file_input'
 module WidgetTests
   class Test_FileInput_widget < BureaucratTestCase
     def test_correctly_render
-      input = Widgets::FileInput.new
+      input = Bureaucrat::Widgets::FileInput.new
       expected = normalize_html("<input name='test' type='file'/>")
       rendered = normalize_html(input.render('test', "anything"))
       assert_equal(expected, rendered)
