+++ test/widgets/text_input_test.rb
@@ -4,7 +4,7 @@ module Widget
   class Test_TextInput_widget < BureaucratTestCase
     class Test_with_empty_attributes < BureaucratTestCase
       def test_correctly_render
-        input = Widgets::TextInput.new
+        input = Bureaucrat::Widgets::TextInput.new
         expected = normalize_html('<input type="text" value="hi" name="test" />')
         rendered = normalize_html(input.render('test', 'hi'))
         assert_equal(expected, rendered)
@@ -13,7 +13,7 @@ module Widget
 
     class Test_with_attributes < BureaucratTestCase
       def test_correctly_render
-        input = Widgets::TextInput.new(attribute: 'value')
+        input = Bureaucrat::Widgets::TextInput.new(attribute: 'value')
         expected = normalize_html('<input type="text" value="hi" name="test" attribute="value" />')
         rendered = normalize_html(input.render('test', 'hi'))
         assert_equal(expected, rendered)
@@ -22,7 +22,7 @@ module Widget
 
     class Test_without_value < BureaucratTestCase
       def test_not_render_a_value
-        input = Widgets::TextInput.new
+        input = Bureaucrat::Widgets::TextInput.new
         expected = normalize_html('<input type="text" name="test" />')
         rendered = normalize_html(input.render('test', nil))
         assert_equal(expected, rendered)
@@ -31,7 +31,7 @@ module Widget
 
     class Test_when_copied < BureaucratTestCase
       def test_have_a_copy_of_the_attributes
-        input1 = Widgets::TextInput.new(attribute: 2)
+        input1 = Bureaucrat::Widgets::TextInput.new(attribute: 2)
         input2 = input1.dup
         assert_not_equal(input1.attrs.object_id, input2.attrs.object_id)
       end
